Added options and error handling to init-game

+++ src/initialization/init-game.js
@@ -1,7 +1,24 @@
-// Initialize crops list
+// Handle arguments to init-game
+const arguments = process.argv.slice(2);
+let deleteFlag = false;
 
+if (arguments.length === 0) console.log('Running default setup');
+
+if (arguments.includes('-h')) {
+    const { content } = require('./help.json');
+    console.log(content);
+    process.exit(0);
+}
+
+if (arguments.includes('-d')) {
+    deleteFlag = true;
+    console.log("Delete Flag: True");
+}
+
+// Database
 const Crop = require('../models/crop.model.js');
-const cropList = require('../data/crops.json');
+const Player = require('../models/player.model.js');
+const cropList = require('../../data/crops.json');
 
 const mongoose = require('mongoose');
 require('dotenv').config();
@@ -16,11 +33,15 @@ db.on('error', console.error.bind(console, 'connection to mongodb error'));
 db.once('open', async () => {
     console.log('Connection to mongoDB server Success');
 
-    await Crop.deleteMany({});
-    console.log('Deleted all crops');
+    if (deleteFlag) {
+        await Crop.deleteMany({});
+        console.log('Deleted all crops');
+        await Player.deleteMany({});
+        console.log('Deleted all players');
+    }
 
     const promises = [];
-
+    
     cropList.forEach(crop => {
         promises.push(addCrop(crop));
     })
@@ -34,6 +55,13 @@ db.once('open', async () => {
 });
 
 async function addCrop(cropObj) {
-    await Crop.create(cropObj);
-    console.log(`Added ${cropObj.name}`);
+    try {
+        await Crop.create(cropObj);
+        console.log(`Added ${cropObj.name}`);
+    } catch (error) {
+        if (error.name === 'MongoServerError' && error.code === 11000)
+            console.log(`Already found in database: ${cropObj.name}`)
+        else
+            console.log(error);
+    }
 }
